This original script has been replaced by the codes from humboldt70_custom.R that I have been working with.
This updated script does not require tcltk package (and therefore the ks package).

One plot (*_SP-OVERLAP.pdf) from humboldt.doitall could not be made due to the requirement of the ks package. However, this was not necessary and can be plotted post-hoc locally if need be.

PNTI calculations were hashed off as the values appear unreliable in this case. Likely due to the nature of the environmental variables (bounded).

Parts of the codes which calls the functions from the installed humboldt package (through humboldt:: or library(humboldt)) were altered so that the functions could be called from this script instead.